Shuffling files around to keep symbols under the MapValidator namespace
