--- lib/vocabularies.rb
@@ -1,86 +1,0 @@
-def getVocabularies
-  {
-    dateQualifierVocabularies: [
-      '[blank]',
-      'Approximate'
-    ],
-    formatPhysical: [
-      'Linen',
-      'Map',
-      'Magnetic media',
-      'Microform',
-      'Motion Picture',
-      'Negative, slide, or transparency',
-      'Object',
-      'Optical Media',
-      'Paper',
-      'Plans',
-      'Phonographic media',
-      'Photographic print',
-      'Plastic film',
-      'Vellum or Parchment',
-      'Volume'
-    ],
-    containedWith: [
-      '[blank]',
-      'Digital - Cartridge Tape (LTO)',
-      'Digital - Compact Disc (CD)',
-      'Digital - Digital Versatile Disc (DVD)',
-      'Digital - Floppy Disk',
-      'Digital - Hard Disk Drive',
-      'Digital - Network Attached Storage (NAS)',
-      'Digital - Universal Serial Bus (USB)',
-      'Physical - Album',
-      'Physical - Archives Box',
-      'Physical - Bundle',
-      'Physical - Card box',
-      'Physical - Card Drawer',
-      'Physical - Film canister',
-      'Physical - Microfilm box',
-      'Physical - Other',
-      'Physical - Phase Box',
-      'Physical - Slide carousel',
-      'Physical - Solander Box',
-      'Physical - Type 1 box',
-      'Physical - Type 10 box',
-      'Physical - Type 11 box',
-      'Physical - Type 2 box',
-      'Physical - Type 3.1 box',
-      'Physical - Type 5.1 box',
-      'Physical - Type 6 box'
-    ],
-    accessCategories: [
-      '0',
-      '5',
-      '10',
-      '15',
-      '20',
-      '25',
-      '30',
-      '35',
-      '40',
-      '45',
-      '50',
-      '55',
-      '60',
-      '65',
-      '70',
-      '75',
-      '80',
-      '85',
-      '90',
-      '95',
-      '100',
-      'Overriding Legislation',
-      'Non Publication Order',
-      'Sealed By Court'
-    ],
-    sensitivityLabels: [
-      'Cultural Sensitivity Warning',
-      'Public Access/Published',
-      'Secret and Sacred',
-      'These records may contain terms or images which Aboriginal and Torres Strait Islander people may find distressing or offensive',
-      'This record may contain information or photographs which some people may find distressing'
-    ]
-  }
-end
