+++ tests/validator_functions.test.rb
@@ -1,11 +1,13 @@
 require 'test/unit'
-require_relative '../lib/validator_functions'
-require_relative '../lib/notifications'
+require_relative '../lib/map_validator/validator_functions'
+require_relative '../lib/map_validator/notifications'
 
 class TestNotifications < Test::Unit::TestCase
 
+  include MapValidator::ValidatorFunctions
+
   def setup
-    @notifications = Notifications.new
+    @notifications = MapValidator::Notifications.new
     @meta = { col: '10', row: 'Test' }
   end
